Added home items counter
